feat: implement Milestone Basic Tests for product milestone tracking and retrieval
